backend/fastapi-api/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.39.1
distro==1.9.0
dnspython==2.6.1
exceptiongroup==1.2.0
fastapi==0.110.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
openai==1.17.1
packaging==24.0
psycopg2-binary==2.9.9
pydantic==2.7.0
pydantic_core==2.18.1
pymongo==4.6.3
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
